requirements.txt
dash==2.14.1
plotly==5.17.0
pandas==2.0.3
pyodbc==4.0.37
python-dotenv==1.0.0
gunicorn==21.2.0
